+++ tests/ExponentialBackoffTest.php
@@ -4,8 +4,6 @@ declare(strict_types = 1);
 namespace Tests\Innmind\HttpTransport;
 
 use Innmind\HttpTransport\{
-    ExponentialBackoff,
-    Curl,
     Transport,
     ServerError,
     Success,
@@ -23,9 +21,8 @@ use Innmind\Http\{
     ProtocolVersion,
     Response\StatusCode,
 };
-use Innmind\TimeWarp\Halt;
-use Innmind\TimeContinuum\{
-    Clock,
+use Innmind\Time\{
+    Halt,
     Period,
 };
 use Innmind\Url\Url;
@@ -38,17 +35,6 @@ use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class ExponentialBackoffTest extends TestCase
 {
-    public function testInterface()
-    {
-        $this->assertInstanceOf(
-            Transport::class,
-            ExponentialBackoff::of(
-                Curl::of(Clock::live()),
-                Halt\Usleep::new(),
-            ),
-        );
-    }
-
     public function testDoesntRetryWhenInformationResponseOnFirstCall()
     {
         $request = Request::of(
@@ -62,24 +48,9 @@ class ExponentialBackoffTest extends TestCase
         );
         $expected = Either::left(new Information($request, $response));
 
-        $fulfill = ExponentialBackoff::of(
-            new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                ) {
-                }
-
-                public function __invoke(Request $request): Either
-                {
-                    return $this->expected;
-                }
-            },
-            new class implements Halt {
-                public function __invoke(Period $period): Attempt
-                {
-                    return Attempt::error(new \Exception);
-                }
-            },
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static fn() => $expected),
+            Halt::via(static fn() => Attempt::error(new \Exception)),
         );
 
         $this->assertEquals($expected, $fulfill($request));
@@ -98,24 +69,9 @@ class ExponentialBackoffTest extends TestCase
         );
         $expected = Either::right(new Success($request, $response));
 
-        $fulfill = ExponentialBackoff::of(
-            new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                ) {
-                }
-
-                public function __invoke(Request $request): Either
-                {
-                    return $this->expected;
-                }
-            },
-            new class implements Halt {
-                public function __invoke(Period $period): Attempt
-                {
-                    return Attempt::error(new \Exception);
-                }
-            },
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static fn() => $expected),
+            Halt::via(static fn() => Attempt::error(new \Exception)),
         );
 
         $this->assertEquals($expected, $fulfill($request));
@@ -134,24 +90,9 @@ class ExponentialBackoffTest extends TestCase
         );
         $expected = Either::left(new Redirection($request, $response));
 
-        $fulfill = ExponentialBackoff::of(
-            new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                ) {
-                }
-
-                public function __invoke(Request $request): Either
-                {
-                    return $this->expected;
-                }
-            },
-            new class implements Halt {
-                public function __invoke(Period $period): Attempt
-                {
-                    return Attempt::error(new \Exception);
-                }
-            },
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static fn() => $expected),
+            Halt::via(static fn() => Attempt::error(new \Exception)),
         );
 
         $this->assertEquals($expected, $fulfill($request));
@@ -170,24 +111,9 @@ class ExponentialBackoffTest extends TestCase
         );
         $expected = Either::left(new ClientError($request, $response));
 
-        $fulfill = ExponentialBackoff::of(
-            new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                ) {
-                }
-
-                public function __invoke(Request $request): Either
-                {
-                    return $this->expected;
-                }
-            },
-            new class implements Halt {
-                public function __invoke(Period $period): Attempt
-                {
-                    return Attempt::error(new \Exception);
-                }
-            },
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static fn() => $expected),
+            Halt::via(static fn() => Attempt::error(new \Exception)),
         );
 
         $this->assertEquals($expected, $fulfill($request));
@@ -202,24 +128,9 @@ class ExponentialBackoffTest extends TestCase
         );
         $expected = Either::left(new MalformedResponse($request));
 
-        $fulfill = ExponentialBackoff::of(
-            new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                ) {
-                }
-
-                public function __invoke(Request $request): Either
-                {
-                    return $this->expected;
-                }
-            },
-            new class implements Halt {
-                public function __invoke(Period $period): Attempt
-                {
-                    return Attempt::error(new \Exception);
-                }
-            },
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static fn() => $expected),
+            Halt::via(static fn() => Attempt::error(new \Exception)),
         );
 
         $this->assertEquals($expected, $fulfill($request));
@@ -234,24 +145,9 @@ class ExponentialBackoffTest extends TestCase
         );
         $expected = Either::left(new Failure($request, 'whatever'));
 
-        $fulfill = ExponentialBackoff::of(
-            new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                ) {
-                }
-
-                public function __invoke(Request $request): Either
-                {
-                    return $this->expected;
-                }
-            },
-            new class implements Halt {
-                public function __invoke(Period $period): Attempt
-                {
-                    return Attempt::error(new \Exception);
-                }
-            },
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static fn() => $expected),
+            Halt::via(static fn() => Attempt::error(new \Exception)),
         );
 
         $this->assertEquals($expected, $fulfill($request));
@@ -269,50 +165,34 @@ class ExponentialBackoffTest extends TestCase
             $request->protocolVersion(),
         );
         $expected = Either::left(new ClientError($request, $response));
+        $calls = 0;
 
-        $fulfill = ExponentialBackoff::of(
-            $inner = new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                    public int $calls = 0,
-                ) {
-                }
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static function() use (&$calls, $expected) {
+                ++$calls;
 
-                public function __invoke(Request $request): Either
-                {
-                    ++$this->calls;
+                return $expected;
+            }),
+            Halt::via(function($period) {
+                static $calls = 0;
+                ++$calls;
 
-                    return $this->expected;
-                }
-            },
-            new class($this) implements Halt {
-                public function __construct(
-                    private $test,
-                    private int $calls = 0,
-                ) {
-                }
+                match ($calls) {
+                    1, 6 => $this->assertEquals(Period::millisecond(100), $period),
+                    2, 7 => $this->assertEquals(Period::millisecond(271), $period),
+                    3, 8 => $this->assertEquals(Period::millisecond(738), $period),
+                    4, 9 => $this->assertEquals(Period::millisecond(2008), $period),
+                    5, 10 => $this->assertEquals(Period::millisecond(5459), $period),
+                };
 
-                public function __invoke(Period $period): Attempt
-                {
-                    ++$this->calls;
-
-                    match ($this->calls) {
-                        1, 6 => $this->test->assertEquals(Period::millisecond(100), $period),
-                        2, 7 => $this->test->assertEquals(Period::millisecond(271), $period),
-                        3, 8 => $this->test->assertEquals(Period::millisecond(738), $period),
-                        4, 9 => $this->test->assertEquals(Period::millisecond(2008), $period),
-                        5, 10 => $this->test->assertEquals(Period::millisecond(5459), $period),
-                    };
-
-                    return Attempt::result(SideEffect::identity());
-                }
-            },
+                return Attempt::result(SideEffect::identity);
+            }),
         );
 
         $this->assertEquals($expected, $fulfill($request));
         // to make sure halt periods are kept between requests
         $this->assertEquals($expected, $fulfill($request));
-        $this->assertSame(12, $inner->calls);
+        $this->assertSame(12, $calls);
     }
 
     public function testRetryWhileThereIsStillAServerError()
@@ -327,50 +207,34 @@ class ExponentialBackoffTest extends TestCase
             $request->protocolVersion(),
         );
         $expected = Either::left(new ServerError($request, $response));
+        $calls = 0;
 
-        $fulfill = ExponentialBackoff::of(
-            $inner = new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                    public int $calls = 0,
-                ) {
-                }
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static function() use (&$calls, $expected) {
+                ++$calls;
 
-                public function __invoke(Request $request): Either
-                {
-                    ++$this->calls;
+                return $expected;
+            }),
+            Halt::via(function($period) {
+                static $calls = 0;
+                ++$calls;
 
-                    return $this->expected;
-                }
-            },
-            new class($this) implements Halt {
-                public function __construct(
-                    private $test,
-                    private int $calls = 0,
-                ) {
-                }
+                match ($calls) {
+                    1, 6 => $this->assertEquals(Period::millisecond(100), $period),
+                    2, 7 => $this->assertEquals(Period::millisecond(271), $period),
+                    3, 8 => $this->assertEquals(Period::millisecond(738), $period),
+                    4, 9 => $this->assertEquals(Period::millisecond(2008), $period),
+                    5, 10 => $this->assertEquals(Period::millisecond(5459), $period),
+                };
 
-                public function __invoke(Period $period): Attempt
-                {
-                    ++$this->calls;
-
-                    match ($this->calls) {
-                        1, 6 => $this->test->assertEquals(Period::millisecond(100), $period),
-                        2, 7 => $this->test->assertEquals(Period::millisecond(271), $period),
-                        3, 8 => $this->test->assertEquals(Period::millisecond(738), $period),
-                        4, 9 => $this->test->assertEquals(Period::millisecond(2008), $period),
-                        5, 10 => $this->test->assertEquals(Period::millisecond(5459), $period),
-                    };
-
-                    return Attempt::result(SideEffect::identity());
-                }
-            },
+                return Attempt::result(SideEffect::identity);
+            }),
         );
 
         $this->assertEquals($expected, $fulfill($request));
         // to make sure halt periods are kept between requests
         $this->assertEquals($expected, $fulfill($request));
-        $this->assertSame(12, $inner->calls);
+        $this->assertSame(12, $calls);
     }
 
     public function testRetryWhileThereIsStillAConnectionFailure()
@@ -381,50 +245,34 @@ class ExponentialBackoffTest extends TestCase
             ProtocolVersion::v11,
         );
         $expected = Either::left(new ConnectionFailed($request, ''));
+        $calls = 0;
 
-        $fulfill = ExponentialBackoff::of(
-            $inner = new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                    public int $calls = 0,
-                ) {
-                }
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static function() use (&$calls, $expected) {
+                ++$calls;
 
-                public function __invoke(Request $request): Either
-                {
-                    ++$this->calls;
+                return $expected;
+            }),
+            Halt::via(function($period) {
+                static $calls = 0;
+                ++$calls;
 
-                    return $this->expected;
-                }
-            },
-            new class($this) implements Halt {
-                public function __construct(
-                    private $test,
-                    private int $calls = 0,
-                ) {
-                }
+                match ($calls) {
+                    1, 6 => $this->assertEquals(Period::millisecond(100), $period),
+                    2, 7 => $this->assertEquals(Period::millisecond(271), $period),
+                    3, 8 => $this->assertEquals(Period::millisecond(738), $period),
+                    4, 9 => $this->assertEquals(Period::millisecond(2008), $period),
+                    5, 10 => $this->assertEquals(Period::millisecond(5459), $period),
+                };
 
-                public function __invoke(Period $period): Attempt
-                {
-                    ++$this->calls;
-
-                    match ($this->calls) {
-                        1, 6 => $this->test->assertEquals(Period::millisecond(100), $period),
-                        2, 7 => $this->test->assertEquals(Period::millisecond(271), $period),
-                        3, 8 => $this->test->assertEquals(Period::millisecond(738), $period),
-                        4, 9 => $this->test->assertEquals(Period::millisecond(2008), $period),
-                        5, 10 => $this->test->assertEquals(Period::millisecond(5459), $period),
-                    };
-
-                    return Attempt::result(SideEffect::identity());
-                }
-            },
+                return Attempt::result(SideEffect::identity);
+            }),
         );
 
         $this->assertEquals($expected, $fulfill($request));
         // to make sure halt periods are kept between requests
         $this->assertEquals($expected, $fulfill($request));
-        $this->assertSame(12, $inner->calls);
+        $this->assertSame(12, $calls);
     }
 
     public function testStopRetryingWhenNoLongerReceivingAServerError()
@@ -444,44 +292,29 @@ class ExponentialBackoffTest extends TestCase
         );
         $error = Either::left(new ServerError($request, $response1));
         $expected = Either::right(new Success($request, $response2));
+        $all = [$error, $expected];
+        $calls = 0;
 
-        $fulfill = ExponentialBackoff::of(
-            $inner = new class([$error, $expected]) implements Transport {
-                public function __construct(
-                    private $expected,
-                    public int $calls = 0,
-                ) {
-                }
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static function() use (&$calls, &$all) {
+                ++$calls;
 
-                public function __invoke(Request $request): Either
-                {
-                    ++$this->calls;
+                return \array_shift($all);
+            }),
+            Halt::via(function($period) {
+                static $calls = 0;
+                ++$calls;
 
-                    return \array_shift($this->expected);
-                }
-            },
-            new class($this) implements Halt {
-                public function __construct(
-                    private $test,
-                    private int $calls = 0,
-                ) {
-                }
+                match ($calls) {
+                    1 => $this->assertEquals(Period::millisecond(100), $period),
+                };
 
-                public function __invoke(Period $period): Attempt
-                {
-                    ++$this->calls;
-
-                    match ($this->calls) {
-                        1 => $this->test->assertEquals(Period::millisecond(100), $period),
-                    };
-
-                    return Attempt::result(SideEffect::identity());
-                }
-            },
+                return Attempt::result(SideEffect::identity);
+            }),
         );
 
         $this->assertEquals($expected, $fulfill($request));
-        $this->assertSame(2, $inner->calls);
+        $this->assertSame(2, $calls);
     }
 
     public function testByDefaultRetriesFiveTimesByUsingAPowerOfE()
@@ -496,47 +329,31 @@ class ExponentialBackoffTest extends TestCase
             $request->protocolVersion(),
         );
         $expected = Either::left(new ServerError($request, $response));
+        $calls = 0;
 
-        $fulfill = ExponentialBackoff::of(
-            $inner = new class($expected) implements Transport {
-                public function __construct(
-                    private $expected,
-                    public int $calls = 0,
-                ) {
-                }
+        $fulfill = Transport::exponentialBackoff(
+            Transport::via(static function() use (&$calls, $expected) {
+                ++$calls;
 
-                public function __invoke(Request $request): Either
-                {
-                    ++$this->calls;
+                return $expected;
+            }),
+            Halt::via(function($period) {
+                static $calls = 0;
+                ++$calls;
 
-                    return $this->expected;
-                }
-            },
-            new class($this) implements Halt {
-                public function __construct(
-                    private $test,
-                    private int $calls = 0,
-                ) {
-                }
+                match ($calls) {
+                    1=> $this->assertEquals(Period::millisecond(100), $period),
+                    2=> $this->assertEquals(Period::millisecond(271), $period),
+                    3=> $this->assertEquals(Period::millisecond(738), $period),
+                    4=> $this->assertEquals(Period::millisecond(2008), $period),
+                    5=> $this->assertEquals(Period::millisecond(5459), $period),
+                };
 
-                public function __invoke(Period $period): Attempt
-                {
-                    ++$this->calls;
-
-                    match ($this->calls) {
-                        1 => $this->test->assertEquals(Period::millisecond(100), $period),
-                        2 => $this->test->assertEquals(Period::millisecond(271), $period),
-                        3 => $this->test->assertEquals(Period::millisecond(738), $period),
-                        4 => $this->test->assertEquals(Period::millisecond(2008), $period),
-                        5 => $this->test->assertEquals(Period::millisecond(5459), $period),
-                    };
-
-                    return Attempt::result(SideEffect::identity());
-                }
-            },
+                return Attempt::result(SideEffect::identity);
+            }),
         );
 
         $this->assertEquals($expected, $fulfill($request));
-        $this->assertSame(6, $inner->calls);
+        $this->assertSame(6, $calls);
     }
 }